git-svn-id: svn+ssh://svn.code.sf.net/p/lemur/code/indri/trunk@1257 14020d50-3050-45fe-8e0c-5ed7005ca764

+++ include/indri/MemoryIndexDocListFileIterator.hpp
@@ -47,7 +47,7 @@ namespace indri {
         _alphabetical.clear();
         _alphabetical.reserve( _termData.size() );
 
-        for( int i=0; i<_termData.size(); i++ ) {
+        for( size_t i=0; i<_termData.size(); i++ ) {
           _alphabetical.push_back( _termData[i] );
         }
