+++ src/InferenceNetwork.cpp
@@ -119,7 +119,7 @@ void indri::infnet::InferenceNetwork::_indexChanged( indri::index::Index& index 
   _closeIteratorBound = -1;
 
   // doc iterators
-  for( int i=0; i<_termNames.size(); i++ ) {
+  for( size_t i=0; i<_termNames.size(); i++ ) {
     indri::index::DocListIterator* iterator = index.docListIterator( _termNames[i] );
     if( iterator )
       iterator->startIteration();
@@ -128,7 +128,7 @@ void indri::infnet::InferenceNetwork::_indexChanged( indri::index::Index& index 
   }
 
   // field iterators
-  for( int i=0; i<_fieldNames.size(); i++ ) {
+  for( size_t i=0; i<_fieldNames.size(); i++ ) {
     indri::index::DocExtentListIterator* iterator = index.fieldListIterator( _fieldNames[i] );
     if( iterator )
       iterator->startIteration();
@@ -137,7 +137,7 @@ void indri::infnet::InferenceNetwork::_indexChanged( indri::index::Index& index 
   }
   
   // prior iterators
-  for( int i=0; i<_priorNames.size(); i++ ) {
+  for( size_t i=0; i<_priorNames.size(); i++ ) {
     // TODO: this is wasteful, since the prior is associated with the whole collection,
     // there's no need to fetch it for each index.  but, it's just easier to code it like this for now
     indri::collection::PriorListIterator* iterator = _repository.priorListIterator( _priorNames[i] );
@@ -241,17 +241,17 @@ indri::collection::PriorListIterator* indri::infnet::InferenceNetwork::getPriorI
 
 int indri::infnet::InferenceNetwork::addDocIterator( const std::string& termName ) {
   _termNames.push_back( termName );
-  return _termNames.size()-1;
+  return (int)_termNames.size()-1;
 }
 
 int indri::infnet::InferenceNetwork::addFieldIterator( const std::string& fieldName ) {
   _fieldNames.push_back( fieldName );
-  return _fieldNames.size()-1;
+  return (int)_fieldNames.size()-1;
 }
 
 int indri::infnet::InferenceNetwork::addPriorIterator( const std::string& priorName ) {
   _priorNames.push_back( priorName );
-  return _priorNames.size()-1;
+  return (int)_priorNames.size()-1;
 }
 
 void indri::infnet::InferenceNetwork::addListNode( indri::infnet::ListIteratorNode* listNode ) {
@@ -342,7 +342,7 @@ const indri::infnet::InferenceNetwork::MAllResults& indri::infnet::InferenceNetw
   // fetch the current index state
   indri::collection::Repository::index_state indexes = _repository.indexes();
   
-  for( int i=0; i<indexes->size(); i++ ) {
+  for( size_t i=0; i<indexes->size(); i++ ) {
     indri::index::Index& index = *(*indexes)[i];
     indri::thread::ScopedLock iterators( index.iteratorLock() );
