+++ src/InferenceNetworkBuilder.cpp
@@ -282,7 +282,7 @@ static bool inferencenetworkbuilder_contains_stopwords( const std::vector<indri:
 }
 
 static int inferencenetworkbuilder_find_stopwords_left( const std::vector<indri::infnet::ListIteratorNode*>& nodes ) {
-  int begin;
+  size_t begin;
 
   for( begin=0; begin<nodes.size(); begin++ ) {
     if( inferencenetworkbuilder_is_stopword(nodes[begin]) )
@@ -291,7 +291,7 @@ static int inferencenetworkbuilder_find_stopwords_left( const std::vector<indri:
     break;
   }
 
-  return begin;
+  return int(begin);
 }
 
 static int inferencenetworkbuilder_find_stopwords_right( const std::vector<indri::infnet::ListIteratorNode*>& nodes ) {
@@ -300,7 +300,7 @@ static int inferencenetworkbuilder_find_stopwords_right( const std::vector<indri
   if( nodes.size() == 0 )
     return 0;
 
-  for( end=nodes.size()-1; end>=0; end-- ) {
+  for( end=(int)nodes.size()-1; end>=0; end-- ) {
     if( inferencenetworkbuilder_is_stopword(nodes[end]) )
       continue;
 
@@ -1187,11 +1187,11 @@ void indri::infnet::InferenceNetworkBuilder::after( indri::lang::WildcardTerm* w
     // fill the extent or with IndexTerms from the repository
     // that match our wildcard spec...
     indri::collection::Repository::index_state theIndexes = _repository.indexes();
-    int numIndexes=theIndexes->size();
+    size_t numIndexes=theIndexes->size();
 
     std::string normalizedTerm=wildcardTerm->getTerm();
                 
-    for (int i=0; i < numIndexes; i++) {
+    for (size_t i=0; i < numIndexes; i++) {
       // get the index.
       indri::index::Index* thisIndex = (*theIndexes)[i];
